remove unfitting methods form ImageRequest.swift and fix memory issue
