+++ mobile/ios/Runner/Images/ImageRequest.swift
@@ -1,4 +1,3 @@
-import Accelerate
 import Foundation
 
 class ImageRequest {
@@ -18,11 +17,12 @@ class ImageRequest {
   }
 
   func cancel() {
-    guard let cb = state.withLock({
-      $0.isCancelled = true
-      defer { $0.callback = nil }
-      return $0.callback
-    }) else { return }
+    let cb = state.withLock { state in
+      state.isCancelled = true
+      defer { state.callback = nil }
+      return state.callback
+    }
+    guard let cb else { return }
     onCancel()
     cb(ImageProcessing.cancelledResult)
   }
@@ -30,39 +30,16 @@ class ImageRequest {
   /// Delivers the result to the callback. Returns true if the callback was called, false if it was already consumed.
   @discardableResult
   func finish(with result: Result<[String: Int64]?, any Error>) -> Bool {
-    guard let cb = state.withLock({
-      defer { $0.callback = nil }
-      return $0.callback
-    }) else { return false }
+    let cb = state.withLock { state in
+      defer { state.callback = nil }
+      return state.callback
+    }
+    guard let cb else { return false }
     cb(result)
     return true
   }
 
   func onCancel() {}
-
-  func finish(encoding data: Data) {
-    let length = data.count
-    let pointer = malloc(length)!
-    data.copyBytes(to: pointer.assumingMemoryBound(to: UInt8.self), count: length)
-    if !finish(with: .success([
-      "pointer": Int64(Int(bitPattern: pointer)),
-      "length": Int64(length),
-    ])) {
-      free(pointer)
-    }
-  }
-
-  func finish(cgImage: CGImage, format: inout vImage_CGImageFormat) throws {
-    let buffer = try vImage_Buffer(cgImage: cgImage, format: format)
-    if !finish(with: .success([
-      "pointer": Int64(Int(bitPattern: buffer.data)),
-      "width": Int64(buffer.width),
-      "height": Int64(buffer.height),
-      "rowBytes": Int64(buffer.rowBytes),
-    ])) {
-      buffer.free()
-    }
-  }
 }
 
 class RequestRegistry<T: ImageRequest> {
